test:Actualizados mis test de selenium
Se han actualizado los dos test de selenium que tenia asignados y se ha eliminado el archivo selenium.py
Issue #6
